CG-0MP00BL01001RYAJ: Decompose GolfScene into helper classes
Extract responsibilities from 1422-line GolfScene.ts into composable helpers:
- GolfConstants.ts: shared layout, timing, and audio constants
- GolfRenderer.ts: board layout, sprite creation, and refresh logic
- GolfAnimator.ts: card animations, tweens, and drawn-card display
- GolfTurnController.ts: human turn execution and turn flow
- GolfAiController.ts: AI opponent turn execution
- GolfOverlayManager.ts: end-of-game screen overlay
- GolfReplayController.ts: replay mode state injection and takeover

GolfScene.ts reduced from 1422 lines to 446 lines.
All existing tests pass without modification.
